Apply a number of code cleanup refactorings

+++ server/src/test-fast/java/com/thoughtworks/go/server/websocket/ConsoleLogSenderTest.java
@@ -216,7 +216,7 @@ public class ConsoleLogSenderTest {
 
         @Override
         public long totalLinesConsumed() {
-            return (long) count;
+            return count;
         }
 
         @Override
